pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.1.2"
djangorestframework = "^3.10"
djangorestframework_simplejwt = "^4.3"
django-filter = "^2.2"
django-environ = "^0.4.5"
django-constance = {version = "^2.4", extras = ["database"]}
django-extensions = "^2.2"
django-loginas = "^0.3.6"
django-autocompletefilter = "^0.0.6"
django-jazzmin = "^2.3.11"
django-cors-headers = "^3.1"
django-health-check = "^3.16.0"
psycopg2-binary = { version = "^2.8.6", optional = true}
gunicorn = "^20.0.4"
uvicorn = "^0.12.2"

[tool.poetry.extras]
pgsql = ["psycopg2"]

[tool.poetry.dev-dependencies]
Werkzeug = "^0.15.5"
ipython = "^7.11"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
